Replace use of READ_COMMAND by READ_COMMAND_REAL
Both functions are undocumented GAP kernel functions. But the former one
will be removed in GAP 4.9.

Note: neither the old nor the new code performs any error checking. The
return value of READ_COMMAND_REAL is a list, whose first entry indicates
success (true) or failure (false), and the second entry is the return
value of the executed command (if there was any; otherwise, this entry
is not bound).