Avoid DeclareGlobalVariable and InstallValue
... just use BindGlobal instead. The problem with these two functions is that
they rely on GAP "swapping out" the content of a placeholder object (created
by DeclareGlobalVariable) by some other object (via InstallValue). This relies
on some low-level shenanigans in the GAP kernel, and for certain kinds of
values (such as types, families, ...) is known to cause deep problems. While
this is not the case here, we are not generally trying to get away from
InstallValue as much as possible, and in the future may only support it for a
limited subset of values (namely plain data, such as integers, lists, strings
and records).

In any case, just avoiding it usually is trivial, as is the case here.